s3_manager/requirements.txt
flask
jinja2
python-dotenv
gunicorn
boto3
pandas
numpy
xhtml2pdf
requests